refactor: create popup tooltip component
