+++ src/script/components/Popup/index.ts
@@ -1,0 +1,4 @@
+import Popup from './Popup';
+import PopupTooltip from './PopupTooltip';
+
+export { Popup, PopupTooltip };
