+++ src/script/components/SideBar/ChatItem.tsx
@@ -1,9 +1,9 @@
 import { NavLink } from 'react-router-dom';
 import { PageRoutes } from '../../constants';
 import { useState } from 'react';
-import { usePopper } from 'react-popper';
 import { createPortal } from 'react-dom';
 import DeletePopper from './DeletePopper';
+import { PopupTooltip } from '../Popup';
 
 function ChatItem({ id, name, deleteChatItem, editChatItem }: {
   id: number;
@@ -32,11 +32,6 @@ function ChatItem({ id, name, deleteChatItem, editChatItem }: {
   };
 
   const [referenceElement, setReferenceElement] = useState<HTMLButtonElement | null>(null);
-  const [popperElement, setPopperElement] = useState<HTMLDivElement | null>(null);
-
-  const { styles, attributes } = usePopper(referenceElement, popperElement, {
-    placement: 'bottom-end',
-  });
 
   return (
     <li key={id} className='chats__item'>
@@ -46,24 +41,16 @@ function ChatItem({ id, name, deleteChatItem, editChatItem }: {
           <span className='chats__link-btn-icon'></span>
         </button>
         {isVisible && (
-          <>
-            <div className='overlay overlay-tooltip' onClick={hideMore} />
-            <div
-              ref={setPopperElement}
-              style={{ ...styles.popper}}
-              className='popup popup-tooltip'
-              {...attributes.popper}
-            >
-              <button className='popup__btn' onClick={() => editChatItem(id)}>
-                <span className='popup__btn-icon popup__btn-edit'></span>
-                <span>Edit</span>
-              </button>
-              <button className='popup__btn' onClick={openDeleting}>
-                <span className='popup__btn-icon popup__btn-delete'></span>
-                <span>Delete</span>
-              </button>
-            </div>
-          </>
+          <PopupTooltip close={hideMore} refElem={referenceElement}>
+            <button className='popup__btn' onClick={() => editChatItem(id)}>
+              <span className='popup__btn-icon popup__btn-edit'></span>
+              <span>Edit</span>
+            </button>
+            <button className='popup__btn' onClick={openDeleting}>
+              <span className='popup__btn-icon popup__btn-delete'></span>
+              <span>Delete</span>
+            </button>
+          </PopupTooltip>
         )}
         {isDeleting && (
           createPortal(
