+++ src/script/models/index.ts
@@ -1,5 +1,5 @@
 import { AuthResponseError, AuthState } from './auth.model';
 import { FeedItem } from './feed.model';
-import { CreateChatForm, PopupProps } from './popup.model';
+import { CreateChatForm, PopupProps, PopupTooltipProps } from './popup.model';
 
-export type { AuthState, AuthResponseError, CreateChatForm, FeedItem, PopupProps };
+export type { AuthState, AuthResponseError, CreateChatForm, FeedItem, PopupProps, PopupTooltipProps };
